% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-check.R
\name{create_check_pattern}
\alias{create_check_pattern}
\title{Create checkerboard svg pattern}
\usage{
create_check_pattern(
  id,
  angle = 45,
  spacing = 20,
  alpha = 1,
  fg_alpha = 1,
  colour = "#000000",
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{angle}{rotation angle (degrees)}

\item{spacing}{space between features}

\item{alpha}{default: 1.0 (opaque)}

\item{fg_alpha}{alpha of the features}

\item{colour}{colour of the features. e.g. '#345678'}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create checkerboard svg pattern
}
