% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-gradient.R
\name{create_gradient_pattern}
\alias{create_gradient_pattern}
\title{Create linear gradient svg pattern}
\usage{
create_gradient_pattern(
  id,
  angle = 45,
  colour1 = "#ffffff",
  colour2 = "#000000",
  alpha = 1,
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{angle}{rotation angle (degrees)}

\item{colour1, colour2}{the start and end colours of the gradient}

\item{alpha}{default: 1.0 (opaque)}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create linear gradient svg pattern
}
