% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-hex.R
\name{create_hex_pattern}
\alias{create_hex_pattern}
\title{Create hex svg pattern}
\usage{
create_hex_pattern(
  id,
  angle = 0,
  spacing = 20,
  fill_fraction = 0.01,
  alpha = 1,
  fg_alpha = 1,
  colour = "#000000",
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{angle}{rotation angle (degrees)}

\item{spacing}{space between features}

\item{fill_fraction}{size of features. In range [0, 1]}

\item{alpha}{default: 1.0 (opaque)}

\item{fg_alpha}{alpha of the features}

\item{colour}{colour of the features. e.g. '#345678'}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create hex svg pattern
}
