% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-check.R
\name{create_pattern_check}
\alias{create_pattern_check}
\title{Create an SVG checkerboard pattern}
\usage{
create_pattern_check(
  id,
  angle = 45,
  spacing = 20,
  alpha = 1,
  fg_alpha = 1,
  colour = "#000000",
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{angle}{rotation angle (degrees)}

\item{spacing}{space between features}

\item{alpha}{default: 1.0 (opaque)}

\item{fg_alpha}{alpha of the features}

\item{colour}{colour of the features. e.g. '#345678'}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create an SVG pattern which will \code{fill} an element with checks.
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the pattern and add to the SVG definitions
my_pattern <- create_pattern_check(id = 'mypattern')
doc$defs(my_pattern)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  stroke = 'black',
  fill   = my_pattern
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
