% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-gradient.R
\name{create_pattern_gradient}
\alias{create_pattern_gradient}
\title{Create an SVG gradient to use as a pattern}
\usage{
create_pattern_gradient(
  id,
  angle = 45,
  colour1 = "#ffffff",
  colour2 = "#000000",
  alpha = 1,
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{angle}{rotation angle (degrees)}

\item{colour1, colour2}{the start and end colours of the gradient}

\item{alpha}{default: 1.0 (opaque)}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create an SVG pattern which will \code{fill} an element with a colour gradient.
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the pattern and add to the SVG definitions
my_pattern <- create_pattern_gradient(id = 'mypattern')
doc$defs(my_pattern)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  stroke = 'black',
  fill   = my_pattern
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
