% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-stipple.R
\name{create_stipple_pattern}
\alias{create_stipple_pattern}
\title{Create stipple svg pattern}
\usage{
create_stipple_pattern(
  id,
  spacing = 20,
  fill_fraction = 0.2,
  alpha = 1,
  colour = "#000000",
  fg_alpha = 1,
  ...
)
}
\arguments{
\item{id}{id to use for pattern. If NULL then generate rando ID}

\item{spacing}{space between features}

\item{fill_fraction}{size of features. In range [0, 1]}

\item{alpha}{default: 1.0 (opaque)}

\item{colour}{colour of the features. e.g. '#345678'}

\item{fg_alpha}{alpha of the features}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGPattern object
}
\description{
Create stipple svg pattern
}
