% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ggplot.R
\name{gen_pattern_simple_pal_func}
\alias{gen_pattern_simple_pal_func}
\title{Create a pattern palette generator}
\usage{
gen_pattern_simple_pal_func(
  pattern_name = c("stripe", "dot", "hatch", "check"),
  angle = c(22, 45, 67.5, 135),
  spacing = seq(5, 100, length.out = 7),
  fill_fraction = seq(0.1, 0.9, length.out = 3)
)
}
\arguments{
\item{pattern_name}{the name of the pattern. one of 'stripe', 'dot', 'hatch'}

\item{angle}{the vector of angles to choose from (order matters)}

\item{spacing}{the vector of spacing choices}

\item{fill_fraction}{the vector of fill_fraction choices}
}
\value{
a function suitable for a continuous or discrete colour palette generation
}
\description{
Create a pattern palette generator
}
