% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgpatternsimple.R
\docType{package}
\name{svgpatternsimple}
\alias{svgpatternsimple}
\title{svgpatternsimple: A package of parameterised SVG patterns built using \code{minisvg}}
\description{
svgpatternsimple: A package of parameterised SVG patterns built using \code{minisvg}
}
