% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgpatternsimple.R
\docType{package}
\name{svgpatternsimple}
\alias{svgpatternsimple-package}
\alias{svgpatternsimple}
\title{svgpatternsimple: A package of parameterised SVG patterns built using \code{minisvg}}
\description{
Generate simple SVG patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coolbutuseless.github.io/package/svgpatternsimple/,https://github.com/coolbutuseless/svgpatternsimple}
}

}
